+++ src/main/java/org/libreoffice/ci/gerrit/buildbot/config/BuildbotConfigProvider.java
@@ -1,96 +1,148 @@
 package org.libreoffice.ci.gerrit.buildbot.config;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.Reader;
+import java.util.List;
+import java.util.Set;
 
 import org.eclipse.jgit.errors.ConfigInvalidException;
-import org.eclipse.jgit.lib.Config;
+import org.eclipse.jgit.storage.file.FileBasedConfig;
+import org.eclipse.jgit.util.FS;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.base.Preconditions;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Lists;
+import com.google.gerrit.common.errors.NoSuchGroupException;
+import com.google.gerrit.reviewdb.client.AccountGroup;
+import com.google.gerrit.server.account.GroupCache;
 import com.google.gerrit.server.config.SitePaths;
+import com.google.gwtorm.server.OrmException;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 
 public class BuildbotConfigProvider implements Provider<BuildbotConfig> {
 
-	static final Logger log = LoggerFactory.getLogger(BuildbotConfigProvider.class);
+	static final Logger log = LoggerFactory
+			.getLogger(BuildbotConfigProvider.class);
+	
+	private final static String SECTION_USER = "user";
+	private static final String KEY_MAIL = "mail";
+
+	private final static String SECTION_LOG = "log";	
+	private static final String KEY_DIRECTORY = "directory";
+	
+	private final static String SECTION_PROJECT = "project";
+	private final static String KEY_BRANCH = "branch";
+	private final static String KEY_TRIGGER = "trigger";
+	private final static String KEY_REVIEWER_GROUP_NAME = "reviewerGroupName";
 	
 	private SitePaths site;
+
+	private GroupCache groupCache;
+
 	@Inject
-	public BuildbotConfigProvider(SitePaths site) {
+	public BuildbotConfigProvider(SitePaths site, GroupCache groupCache) {
 		this.site = site;
+		this.groupCache = groupCache;
 	}
-	
+
 	@Override
 	public BuildbotConfig get() {
 		BuildbotConfig config = new BuildbotConfig();
-        Config cfg = new Config();
-        File configFile = null;
-        String configContent = null;
-        try {
-            configFile = new File(site.etc_dir, "buildbot.config");
-            configContent = read(configFile);
-        } catch (IOException ex) {
-            log.error(ex.getMessage());
-            throw new IllegalStateException(String.format(
-                    "can not find config file: %s",
-                    configFile.getAbsolutePath()), ex);
-        }
+		File file = new File(site.etc_dir, "buildbot.config");
+		FileBasedConfig cfg = new FileBasedConfig(file, FS.DETECTED);
+		if (!cfg.getFile().exists()) {
+			throw new IllegalStateException(String.format(
+					"can not find config file: %s", file.getAbsolutePath()));
+		}
 
-        try {
-            cfg.fromText(configContent);
-        } catch (ConfigInvalidException ex) {
-            log.error(ex.getMessage());
-            throw new IllegalStateException(String.format(
-                    "can not parse config file: %s",
-                    configFile.getAbsolutePath()), ex);
-        }
-        String email = cfg.getString("user", null, "mail");
-        String project = cfg.getString("project", null, "name");
-        String[] branches = cfg.getStringList("project", null, "branch");
-        String strStrategie = cfg.getString("project", null, "trigger");
+		if (cfg.getFile().length() == 0) {
+			throw new IllegalStateException(String.format(
+					"empty config file: %s", file.getAbsolutePath()));
+		}
 
-        Preconditions.checkNotNull(strStrategie, "strategie must not be null");
+		try {
+			cfg.load();
+		} catch (ConfigInvalidException e) {
+			throw new IllegalStateException(String.format(
+					"config file %s is invalid: %s", cfg.getFile(),
+					e.getMessage()), e);
+		} catch (IOException e) {
+			throw new IllegalStateException(String.format("cannot read %s: %s",
+					cfg.getFile(), e.getMessage()), e);
+		}
 
-        String directory = cfg.getString("log", null, "directory");
-        config = new BuildbotConfig();
-        config.setEmail(email);
-        config.setProject(project);
-        config.setBranches(branches);
+		config = new BuildbotConfig();
+		config.setEmail(cfg.getString(SECTION_USER, null, KEY_MAIL));
+		config.setLogDir(cfg.getString(SECTION_LOG, null, KEY_DIRECTORY));
 
-        TriggerStrategie triggerStrategie = TriggerStrategie.valueOf(strStrategie.toUpperCase());
-        Preconditions.checkNotNull(triggerStrategie, String.format("unknown strategie %s", strStrategie));
+		ImmutableList.Builder<BuildbotProject> dest = ImmutableList.builder();
 
-        config.setTriggerStrategie(triggerStrategie);
-        if (triggerStrategie == TriggerStrategie.POSITIVE_REVIEW) {
-            String reviewerGroupName = cfg.getString("project", null, "reviewerGroupName");
-            Preconditions.checkNotNull(reviewerGroupName, "reviewerGroupName must not be null");
-            config.setReviewerGroupName(reviewerGroupName);
-        }
-        config.setLogDir(directory);
-
-        return config;
+		for (BuildbotProject p : allProjects(cfg)) {
+			dest.add(p);
+		}
+		config.setProjects(dest.build());
+		return config;
 	}
 
-    private String read(final File configFile) throws IOException {
-        final Reader r = new InputStreamReader(new FileInputStream(configFile),
-                "UTF-8");
-        try {
-            final StringBuilder buf = new StringBuilder();
-            final char[] tmp = new char[1024];
-            int n;
-            while (0 < (n = r.read(tmp))) {
-                buf.append(tmp, 0, n);
-            }
-            return buf.toString();
-        } finally {
-            r.close();
-        }
-    }
+	private List<BuildbotProject> allProjects(FileBasedConfig cfg) {
+		Set<String> names = cfg.getSubsections(SECTION_PROJECT);
+		List<BuildbotProject> result = Lists.newArrayListWithCapacity(names
+				.size());
+		for (String name : names) {
+			result.add(parseProject(cfg, name));
+		}
+		return result;
+	}
 
+	private BuildbotProject parseProject(FileBasedConfig cfg, String name) {
+		BuildbotProject p = new BuildbotProject(name);
+		String[] branches = cfg.getStringList(SECTION_PROJECT, name, KEY_BRANCH);
+		String strStrategie = cfg.getString(SECTION_PROJECT, name, KEY_TRIGGER);
+
+		Preconditions.checkNotNull(strStrategie, "strategie must not be null");
+
+		p.setBranches(branches);
+
+		TriggerStrategie triggerStrategie = TriggerStrategie
+				.valueOf(strStrategie.toUpperCase());
+		Preconditions.checkNotNull(triggerStrategie,
+				String.format("unknown strategie %s", strStrategie));
+
+		p.setTriggerStrategie(triggerStrategie);
+		if (triggerStrategie == TriggerStrategie.POSITIVE_REVIEW) {
+			String reviewerGroupName = cfg.getString(SECTION_PROJECT, name,
+					KEY_REVIEWER_GROUP_NAME);
+			Preconditions.checkNotNull(reviewerGroupName,
+					"reviewerGroupName must not be null");
+			p.setReviewerGroupId(getReviewerGroupId(reviewerGroupName));
+		}
+		return p;
+	}
+
+	private AccountGroup.UUID getReviewerGroupId(String reviewerGroupName) {
+		Preconditions.checkNotNull(reviewerGroupName,
+				"ReviewerGroup must not be null");
+		try {
+			return findGroup(reviewerGroupName).getGroupUUID();
+		} catch (OrmException e) {
+			throw new IllegalStateException(
+					String.format("Can not retrieve group: %s",
+							reviewerGroupName));
+		} catch (NoSuchGroupException e) {
+			throw new IllegalStateException(String.format(
+					"Group doesn't exist: %s", reviewerGroupName));
+		}
+	}
+
+	private AccountGroup findGroup(final String name) throws OrmException,
+			NoSuchGroupException {
+		final AccountGroup g = groupCache.get(new AccountGroup.NameKey(name));
+		if (g == null) {
+			throw new NoSuchGroupException(name);
+		}
+		return g;
+	}
 }